another cut of big portion code
