+++ lib/fourOhFour.js
@@ -1,15 +1,23 @@
 'use strict'
 
 const FindMyWay = require('find-my-way')
-
+const Context = require('./context')
 const {
   kRoutePrefix,
   kCanSetNotFoundHandler,
-  kFourOhFourLevelInstance
+  kFourOhFourLevelInstance,
+  kReply,
+  kRequest,
+  kContentTypeParser,
+  kBodyLimit,
+  kLogLevel,
+  kFourOhFourContext,
+  kFourOhFour,
+  kMiddlewares,
   // kBodyLimit,
   // kRoutePrefix,
   // kLogLevel,
-  // kHooks,
+  kHooks
   // kContentTypeParser,
   // kReply,
   // kRequest,
@@ -34,13 +42,14 @@ function fourOhFour (fourOhFourFallBack) {
     reply.code(404).send(new Error('Not Found'))
   }
 
+  // TODO: remove duplicated
   function beforeHandlerWarning () {
     if (beforeHandlerWarning.called) return
     beforeHandlerWarning.called = true
     process.emitWarning('The route option `beforeHandler` has been deprecated, use `preHandler` instead')
   }
 
-  function setNotFoundHandler (opts, handler, _setNotFoundHandler) {
+  function setNotFoundHandler (opts, handler, avvio, routeHandler, buildMiddie) {
     const _fastify = this
     const prefix = this[kRoutePrefix] || '/'
 
@@ -84,9 +93,57 @@ function fourOhFour (fourOhFourFallBack) {
     }
 
     this.after((notHandledErr, done) => {
-      _setNotFoundHandler.call(this, prefix, opts, handler)
+      _setNotFoundHandler.call(this, prefix, opts, handler, avvio, routeHandler, buildMiddie)
       done(notHandledErr)
     })
+  }
+
+  function _setNotFoundHandler (prefix, opts, handler, avvio, routeHandler, buildMiddie) {
+    const context = new Context(
+      opts.schema,
+      handler,
+      this[kReply],
+      this[kRequest],
+      this[kContentTypeParser],
+      opts.config || {},
+      this._errorHandler,
+      this[kBodyLimit],
+      this[kLogLevel]
+    )
+
+    avvio.once('preReady', () => {
+      const context = this[kFourOhFourContext]
+
+      const onRequest = this[kHooks].onRequest
+      const preParsing = this[kHooks].preParsing.concat(opts.preParsing || [])
+      const preValidation = this[kHooks].preValidation.concat(opts.preValidation || [])
+      const preSerialization = this[kHooks].preSerialization.concat(opts.preSerialization || [])
+      const preHandler = this[kHooks].preHandler.concat(opts.beforeHandler || opts.preHandler || [])
+      const onSend = this[kHooks].onSend
+      const onError = this[kHooks].onError
+      const onResponse = this[kHooks].onResponse
+
+      context.onRequest = onRequest.length ? onRequest : null
+      context.preParsing = preParsing.length ? preParsing : null
+      context.preValidation = preValidation.length ? preValidation : null
+      context.preSerialization = preSerialization.length ? preSerialization : null
+      context.preHandler = preHandler.length ? preHandler : null
+      context.onSend = onSend.length ? onSend : null
+      context.onError = onError.length ? onError : null
+      context.onResponse = onResponse.length ? onResponse : null
+
+      context._middie = buildMiddie(this[kMiddlewares])
+    })
+
+    if (this[kFourOhFourContext] !== null && prefix === '/') {
+      Object.assign(this[kFourOhFourContext], context) // Replace the default 404 handler
+      return
+    }
+
+    this[kFourOhFourLevelInstance][kFourOhFourContext] = context
+
+    this[kFourOhFour].router.all(prefix + (prefix.endsWith('/') ? '*' : '/*'), routeHandler, context)
+    this[kFourOhFour].router.all(prefix || '/', routeHandler, context)
   }
 }
